+++ packages/terra-avatar/src/terra-dev-site/test/avatar/facility/InvalidImageFacility.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Facility } from '../../../../index';
+
+export default () => <Facility image="invalid-image-url" alt="Facility" id="invalid-image-facility" />;