+++ packages/terra-form-select/src/_Dropdown.jsx
@@ -76,6 +76,7 @@ const Dropdown = ({
     >
       <Hookshot.Content
         {...customProps}
+        disableOnClickOutside
         className={dropdownClasses}
         onResize={onResize}
         onMouseDown={preventDefault}