+++ packages/terra-form-select/tests/jest/SelectField.test.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import SelectField from '../../src/SelectField';
+import intlContexts from './intl-context-setup';
 
 it('should render a default SelectField component', () => {
   const select = (
@@ -12,7 +13,7 @@ it('should render a default SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallow(select)).toMatchSnapshot();
+  expect(shallow(select, intlContexts.shallowContext)).toMatchSnapshot();
 });
 
 it('should render a default SelectField component with attributes spread onto the select', () => {
@@ -26,5 +27,19 @@ it('should render a default SelectField component with attributes spread onto th
     </SelectField>
   );
 
-  expect(shallow(select)).toMatchSnapshot();
+  expect(shallow(select, intlContexts.shallowContext)).toMatchSnapshot();
+});
+
+it('should render a multiple SelectField component that limits the selection count', () => {
+  const select = (
+    <SelectField help="Help Text" label="Label" maxSelectionCount={2} variant="multiple" defaultValue="blue" selectId="select-id" selectAttrs={{ onClick: () => {} }}>
+      <SelectField.Option value="blue" display="Blue" />
+      <SelectField.Option value="green" display="Green" />
+      <SelectField.Option value="purple" display="Purple" />
+      <SelectField.Option value="red" display="Red" />
+      <SelectField.Option value="violet" display="Violet" />
+    </SelectField>
+  );
+
+  expect(shallow(select, intlContexts.shallowContext)).toMatchSnapshot();
 });