+++ packages/terra-paginator/src/ProgressivePaginator.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import ResponsiveElement from 'terra-responsive-element';
-
 import 'terra-base/lib/baseStyles';
+import KeyCode from 'keycode-js';
 import styles from './Paginator.module.scss';
 
-import { calculatePages, KEYCODES } from './_paginationUtils';
+import { calculatePages } from './_paginationUtils';
 
 const cx = classNames.bind(styles);
 
@@ -56,7 +56,7 @@ class ProgressivePaginator extends React.Component {
 
   handleOnKeyDown(index) {
     return (event) => {
-      if (event.nativeEvent.keyCode === KEYCODES.ENTER || event.nativeEvent.keyCode === KEYCODES.SPACE) {
+      if (event.nativeEvent.keyCode === KeyCode.KEY_RETURN || event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
         event.preventDefault();
 
         if (Number.isNaN(Number(index))) {