+++ packages/terra-button/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Replaced local keyCode values with keycode-js npm package 
 
 3.6.0 - (February 26, 2019)
 ------------------