+++ packages/terra-form-select/CHANGELOG.md
@@ -4,7 +4,25 @@ ChangeLog
 Unreleased
 ----------
 ### Fixed
+* Checked for undefined and null for defaultValue
+
+### Added
+* Added new translation strings
+
+### Changed
+* Replaced local keyCode values with keycode-js npm package
+
+5.9.0 - (March 6th, 2019)
+------------------
+### Added
+* Added option to limit the maximum number of selection count for multi-select variants.
+* Introduce an option to allow clearing the currently selected item
+
+5.8.0 - (March 5th, 2019)
+------------------
+### Fixed
 * Race condition where timeout would update after the component is not mounted
+* 'Unable to get property 'getInstance' of undefined or null reference' error in IE
 
 5.7.0 - (February 26, 2019)
 ------------------