+++ packages/terra-status-view/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.7.0 - (March 5, 2019)
+------------------
+### Fixed
+* Update SVG style to `display: block` to fix SVG size and stop text from shifting downards.
+
 3.6.0 - (February 26, 2019)
 ------------------
 ### Added